Move verifier test method
